Added checks for dealer emails for custom content and attachment

+++ app/Observers/DealerEmailObserver.php
@@ -9,7 +9,7 @@ class DealerEmailObserver
 {
     public function updated(DealerEmail $dealerEmail): void
     {
-        if ($dealerEmail->isDirty('attachment')) {
+        if ($dealerEmail->isDirty('attachment') && !is_null($dealerEmail->getOriginal('attachment'))) {
             Storage::disk('public')->delete($dealerEmail->getOriginal('attachment'));
         }
     }
